Chore: abstract validation handlers from `genreFormHandler.js` to separate modules within `/input-handlers/`
